Revert bogus int -> size_t conversions, FATAL on 64bit platforms.
NOTE: size_t is not the same size as int, it happens to be on x86.
However, the compiler will only see "pointer to scalar, fine" and won't
complain - any dereferenced pointer access however is fatal and breaks
on SPARC64.

If flex generates "int yyleng", that's what we must use in
our pointers as well! If we want size_t, we need to have flex fixed!