Forgo DB_NOSYNC when closing transactional DB,
because the original reason for the flag is unclear,
and trim the list of supported Berkeley DB versions
to 5.3. There is no code to deliberately remove support
for other versions, it's just that 5.3 appears to be
the most widespread version these days. Prior versions
can be replaced by 5.3, and 6.x versions have a license
with a stronger copyleft.

If this causes EINVAL to come back in older database versions,
we may want to check that we are not writing to read-only
databases - this was recently found by Steffen Nurpmeso
when adding LMDB support.