Fix out-of-bound read when setting message id
As-is, in set_msg_id(), the code uncoditionally copies 90 bytes from the lex
buffer when setting the message id. Besides being wasteful, this leads to
out-of-bounds reads for certain messages and thus undefined behaviour - think
program termination when accessing a protected page or some noise in a memory
debugger like address sanitizer.

Fixes SourceForce.Net Bug #118.