Fix heap-buffer-overflow in yy_get_next_buffer (SF.net bug #123)
For certain messages with ultra-long tokens, the generated lexer writes
out-of-bounds over the end of a heap allocated buffer.

This is caused by get_decoded_line() modifying the input buffer (and not just
the text), which isn't expected by buf_fill() and thus, yyinput() may return a
count greater than the provided buffer size (cf. the size parameter of
yyinput()).

The attached patch fixes this issue - it changes the following:

 -  add some asserts to yyinput() to detect these issues before entereing flex
    generated internals

 -  eliminate the memcpy() of the input buffer in get_decoded_line() and return
    a special return code instead

 -  adjust all callers of get_decoded_line() to deal with this new return code
    (just 2 locations)

Note that this is just another issue caused by the memcpy() in
get_decoded_line() - for the other one see issue #121.

The attached diff is based on the #121 patch. Patches are against bogofilter
1.2.4 (as packaged with fedora 29).